Cleanup of certificate handling
